changed double to float

+++ cuda_ray_ben.cu
@@ -15,9 +15,9 @@ void circshift(int* out, int* in, int numrows, int numcols, int rowshift, int co
 	}
 }
 /*
-__device__ void columnSum(double* sum, double* mat, int rows, int cols){
+__device__ void columnSum(float* sum, float* mat, int rows, int cols){
 	for(int j = 0; j < cols; j++){
-		double colsum = 0;
+		float colsum = 0;
 		for(int i = 0; i < rows; i++){
 			colsum += mat[i*cols+j];
 		}
@@ -26,7 +26,7 @@ __device__ void columnSum(double* sum, double* mat, int rows, int cols){
 }
 */
 
-void printMat(double* mat, int rows, int cols){
+void printMat(float* mat, int rows, int cols){
         for(int i = 0; i < rows; i++){
                 for(int j = 0; j < cols; j++){
                         std::cout << mat[i*cols+j]<<" ";
@@ -35,7 +35,7 @@ void printMat(double* mat, int rows, int cols){
         }
 	std::cout<<"\n";
 }
-void write_vector_to_file(double* vector, int dim, int evolution) {
+void write_vector_to_file(float* vector, int dim, int evolution) {
 	std::ofstream myfile;
 	myfile.open("evolution"+std::to_string(evolution)+".txt");
 	for(int i = 0; i < dim; i++){
@@ -43,7 +43,7 @@ void write_vector_to_file(double* vector, int dim, int evolution) {
 	}
 }
 
-__global__ void timestep(double* fIn, double* fOut,double* fEq,double* force, double* rho, double* T, double* ux, double* uy, double* tIn, double* tOut, double *tEq, int lx, int ly, int* cxNS, int* cyNS,int* cxT, int* cyT, double* tNS,double* tT, double Thot, int Tcold, double omegaNS, double omegaT, int* oppNS, int* stmNS, int* stmT){
+__global__ void timestep(float* fIn, float* fOut,float* fEq,float* force, float* rho, float* T, float* ux, float* uy, float* tIn, float* tOut, float *tEq, int lx, int ly, int* cxNS, int* cyNS,int* cxT, int* cyT, float* tNS,float* tT, float Thot, int Tcold, float omegaNS, float omegaT, int* oppNS, int* stmNS, int* stmT){
 	int nid = threadIdx.x + blockIdx.x*blockDim.x;
 	int nnodes = lx*ly;
 	if (nid < nnodes){
@@ -63,14 +63,14 @@ __global__ void timestep(double* fIn, double* fOut,double* fEq,double* force, do
 		uy[nid] = (fIn[2*nnodes+nid]-fIn[4*nnodes+nid]+fIn[5*nnodes+nid]+fIn[6*nnodes+nid]-fIn[7*nnodes+nid]-fIn[8*nnodes+nid])/rho[nid];
 		//collision
 		for(int spd = 0; spd < 9; spd++){
-			double cu = 3*(cxNS[spd]*ux[nid] + cyNS[spd]*uy[nid]);
+			float cu = 3*(cxNS[spd]*ux[nid] + cyNS[spd]*uy[nid]);
 			fEq[spd*nnodes + nid] = tNS[spd]*rho[nid]*(1+cu+(0.5)*cu*cu - (1.5)*(ux[nid]*ux[nid]+uy[nid]*uy[nid]));
 			force[spd*nnodes + nid] = 3*tNS[spd]*rho[nid]*(T[nid]-((Thot+Tcold)/2))*(cyNS[spd]*0.001)/(Thot-Tcold);
 			fOut[spd*nnodes + nid] = fIn[spd*nnodes+nid]-omegaNS*(fIn[spd*nnodes+nid]-fEq[spd*nnodes+nid])+force[spd*nnodes+nid];
 		}
 		
 		for(int i = 0; i < 5; i++){
-			double cu = 3*(cxT[i]*ux[nid] + cyT[i]*uy[nid]);
+			float cu = 3*(cxT[i]*ux[nid] + cyT[i]*uy[nid]);
 			tEq[i*nnodes+nid] = T[nid]*tT[i]*(1+cu);
 			tOut[i*nnodes+nid] = tIn[i*nnodes+nid]-omegaT*(tIn[i*nnodes+nid]-tEq[i*nnodes+nid]);
 		}
@@ -109,18 +109,18 @@ int main(int argc, char* argv[]) {
 	int aspect_ratio = 2;
 	int lx = ly*aspect_ratio; //kernel arg
 	int nnodes = lx*ly; //kernel arg
-	double delta_x = 1.0/(ly-2);
-	double Pr = 1.0;
-	double Ra = 1000000;
-	double gr = 0.001;
-	double Thot = 1.0; //kernel arg
+	float delta_x = 1.0/(ly-2);
+	float Pr = 1.0;
+	float Ra = 1000000;
+	float gr = 0.001;
+	float Thot = 1.0; //kernel arg
 	int Tcold = 0; //kernel arg
-	//double T0 = (Thot+Tcold)/2;
-	double delta_t = sqrt(gr*delta_x);
-	double nu = (sqrt(Pr/Ra)*delta_t)/(delta_x*delta_x);
-	double k = sqrt(1.0/(Pr*Ra))*delta_t/(delta_x*delta_x);
-	double omegaNS = 1.0/(3*nu + 0.5);//kernel arg
-	double omegaT = 1.0/(3*k + 0.5);//kernel arg
+	//float T0 = (Thot+Tcold)/2;
+	float delta_t = sqrt(gr*delta_x);
+	float nu = (sqrt(Pr/Ra)*delta_t)/(delta_x*delta_x);
+	float k = sqrt(1.0/(Pr*Ra))*delta_t/(delta_x*delta_x);
+	float omegaNS = 1.0/(3*nu + 0.5);//kernel arg
+	float omegaT = 1.0/(3*k + 0.5);//kernel arg
 	
 	//host variables
 	int maxT = 1000;
@@ -128,36 +128,36 @@ int main(int argc, char* argv[]) {
 	//int Vis_ind = 0;
 	
 	//device needs to know these values
-	double tNS[] = {4./9,1./9,1./9,1./9,1./9,1./36,1./36,1./36,1./36};//kernel arg
+	float tNS[] = {4./9,1./9,1./9,1./9,1./9,1./36,1./36,1./36,1./36};//kernel arg
 	int cxNS[] = {0,1,0,-1,0,1,-1,-1,1};//kernel arg
 	int cyNS[] = {0,0,1,0,-1,1,1,-1,-1};//kernel arg
 	int oppNS[] = {0,3,4,1,2,7,8,5,6};//kernel arg
 
 	
-	double *tNS_d;
+	float *tNS_d;
 	int *cxNS_d, *cyNS_d, *oppNS_d;
-	cudaMalloc((void**)&tNS_d,9*sizeof(double));
+	cudaMalloc((void**)&tNS_d,9*sizeof(float));
 	cudaMalloc((void**)&cxNS_d, 9*sizeof(int));
 	cudaMalloc((void**)&cyNS_d, 9*sizeof(int));
 	cudaMalloc((void**)&oppNS_d, 9*sizeof(int));
-	cudaMemcpy(tNS_d, tNS, 9*sizeof(double),cudaMemcpyHostToDevice);
+	cudaMemcpy(tNS_d, tNS, 9*sizeof(float),cudaMemcpyHostToDevice);
 	cudaMemcpy(cxNS_d, cxNS, 9*sizeof(int),cudaMemcpyHostToDevice);
 	cudaMemcpy(cyNS_d, cyNS, 9*sizeof(int),cudaMemcpyHostToDevice);
 	cudaMemcpy(oppNS_d, oppNS, 9*sizeof(int),cudaMemcpyHostToDevice);
 	
 
-	double tT[] = {1./3,1./6,1./6,1./6,1./6};//kernel arg
+	float tT[] = {1./3,1./6,1./6,1./6,1./6};//kernel arg
 	int cxT[] = {0,1,0,-1,0};//kernel arg
 	int cyT[] = {0,0,1,0,-1};//kernel arg
 	//int oppT[] = {0,3,4,1,2};//kernel arg
 	
-	double *tT_d;
+	float *tT_d;
 	int *cxT_d, *cyT_d;
-	cudaMalloc(&tT_d,5*sizeof(double));
+	cudaMalloc(&tT_d,5*sizeof(float));
 	cudaMalloc(&cxT_d,5*sizeof(int));
 	cudaMalloc(&cyT_d,5*sizeof(int));
 	
-	cudaMemcpy(tT_d, tT, 5*sizeof(double),cudaMemcpyHostToDevice);
+	cudaMemcpy(tT_d, tT, 5*sizeof(float),cudaMemcpyHostToDevice);
 	cudaMemcpy(cxT_d, cxT, 5*sizeof(int),cudaMemcpyHostToDevice);
 	cudaMemcpy(cyT_d, cyT, 5*sizeof(int),cudaMemcpyHostToDevice);
 	
@@ -181,14 +181,14 @@ int main(int argc, char* argv[]) {
 	*/
 
 	//initialize stuff
-	double* fEven = new double [9*nnodes]; //initialize fIn
+	float* fEven = new float [9*nnodes]; //initialize fIn
 	for(int s = 0; s < 9; s++){
 		for(int n = 0; n < nnodes; n++){
 			fEven[s*nnodes+n] = tNS[s];
 		}
 	}
 
-	double* tEven = new double [5*nnodes]; //initialize tIn
+	float* tEven = new float [5*nnodes]; //initialize tIn
 	for(int s = 0; s < 5; s++){
 		for(int n = 0; n < nnodes; n++){
 			tEven[s*nnodes+n] = tT[s]*Tcold;
@@ -239,51 +239,51 @@ int main(int argc, char* argv[]) {
 	cudaMemcpy(stmT_d, stmT, 5*nnodes*sizeof(int), cudaMemcpyHostToDevice);
 	
 	
-	double* T = new double[nnodes];
-	double* ux = new double[nnodes];
-	double* uy = new double[nnodes];
+	float* T = new float[nnodes];
+	float* ux = new float[nnodes];
+	float* uy = new float[nnodes];
 	//here down not actually needed just for debugging purposes
-	double* rho = new double[nnodes];
-	double* fEq = new double[9*nnodes];
-	double* fOut = new double[9*nnodes];
-	double* force = new double[9*nnodes];
-	double* tEq = new double[5*nnodes];
-	double* tOut = new double[5*nnodes];
+	float* rho = new float[nnodes];
+	float* fEq = new float[9*nnodes];
+	float* fOut = new float[9*nnodes];
+	float* force = new float[9*nnodes];
+	float* tEq = new float[5*nnodes];
+	float* tOut = new float[5*nnodes];
 	//here up ''
 
-	double *rho_d, *T_d, *ux_d, *uy_d, *fEven_d, *fOdd_d, *fEq_d, *force_d, *tEven_d, *tOdd_d, *tEq_d;
+	float *rho_d, *T_d, *ux_d, *uy_d, *fEven_d, *fOdd_d, *fEq_d, *force_d, *tEven_d, *tOdd_d, *tEq_d;
 	
-	cudaMalloc(&rho_d, nnodes*sizeof(double));
-	cudaMalloc(&T_d, nnodes*sizeof(double));
-	cudaMalloc(&ux_d, nnodes*sizeof(double));
-	cudaMalloc(&uy_d, nnodes*sizeof(double));
-	cudaMalloc(&fEven_d, 9*nnodes*sizeof(double));
-	cudaMalloc(&fOdd_d, 9*nnodes*sizeof(double));
-	cudaMalloc(&fEq_d, 9*nnodes*sizeof(double));
-	cudaMalloc(&force_d, 9*nnodes*sizeof(double));
-	cudaMalloc(&tEven_d, 5*nnodes*sizeof(double));
-	cudaMalloc(&tOdd_d, 5*nnodes*sizeof(double));
-	cudaMalloc(&tEq_d, 5*nnodes*sizeof(double));
+	cudaMalloc(&rho_d, nnodes*sizeof(float));
+	cudaMalloc(&T_d, nnodes*sizeof(float));
+	cudaMalloc(&ux_d, nnodes*sizeof(float));
+	cudaMalloc(&uy_d, nnodes*sizeof(float));
+	cudaMalloc(&fEven_d, 9*nnodes*sizeof(float));
+	cudaMalloc(&fOdd_d, 9*nnodes*sizeof(float));
+	cudaMalloc(&fEq_d, 9*nnodes*sizeof(float));
+	cudaMalloc(&force_d, 9*nnodes*sizeof(float));
+	cudaMalloc(&tEven_d, 5*nnodes*sizeof(float));
+	cudaMalloc(&tOdd_d, 5*nnodes*sizeof(float));
+	cudaMalloc(&tEq_d, 5*nnodes*sizeof(float));
 
 	
-	cudaMemcpy(fEven_d, fEven, 9*nnodes*sizeof(double), cudaMemcpyHostToDevice);
-	cudaMemcpy(tEven_d, tEven, 5*nnodes*sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(fEven_d, fEven, 9*nnodes*sizeof(float), cudaMemcpyHostToDevice);
+	cudaMemcpy(tEven_d, tEven, 5*nnodes*sizeof(float), cudaMemcpyHostToDevice);
 	
 	for(int cycle = 0; cycle < maxT; cycle++){
 		timestep<<<((nnodes+127)/128),128>>>(fEven_d,fOdd_d,fEq_d,force_d, rho_d, T_d, ux_d, uy_d, tEven_d, tOdd_d, tEq_d, lx, ly, cxNS_d, cyNS_d,cxT_d,cyT_d, tNS_d,tT_d, Thot, Tcold, omegaNS, omegaT, oppNS_d, stmNS_d,stmT_d);
 
 	}
 	//just for debugging
-	cudaMemcpy(T,T_d,nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(ux,ux_d,nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(uy,uy_d,nnodes*sizeof(double),cudaMemcpyDeviceToHost);
+	cudaMemcpy(T,T_d,nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(ux,ux_d,nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(uy,uy_d,nnodes*sizeof(float),cudaMemcpyDeviceToHost);
 	/*
-	cudaMemcpy(rho,rho_d,nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(fEq,fEq_d,9*nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(fOut,fOdd_d,9*nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(force,force_d,9*nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(tEq,tEq_d,5*nnodes*sizeof(double),cudaMemcpyDeviceToHost);
-	cudaMemcpy(tOut,tOdd_d,5*nnodes*sizeof(double),cudaMemcpyDeviceToHost);
+	cudaMemcpy(rho,rho_d,nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(fEq,fEq_d,9*nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(fOut,fOdd_d,9*nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(force,force_d,9*nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(tEq,tEq_d,5*nnodes*sizeof(float),cudaMemcpyDeviceToHost);
+	cudaMemcpy(tOut,tOdd_d,5*nnodes*sizeof(float),cudaMemcpyDeviceToHost);
 	*/
 	
 	printMat(T,1,nnodes);
